requirements.txt
pyspark
selenium
beautifulsoup4
numpy
pandas
python-dotenv
ipykernel
jupyter
html5lib
s3fs